sync

+++ boxes/groups/eos-sdk/hooks-cpp-contracts/extensions/hooks/post-install/02-cpp-contracts.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const { execPromise, emojMap } = require('../../helpers/_exec');
-var tidyChecks = ["*"];
+var tidyChecks = ['*'];
 var path = require('path');
 var CMakeLists = `get_filename_component(PROJ_NAME "\${CMAKE_CURRENT_SOURCE_DIR}" NAME )
 cmake_minimum_required(VERSION 3.5)
@@ -11,17 +11,16 @@ if(EOSIO_CDT_ROOT STREQUAL "" OR NOT EOSIO_CDT_ROOT)
 endif()
 add_contract( \${PROJ_NAME} \${PROJ_NAME} \${PROJ_NAME}.cpp )
 include_directories( \${PROJ_NAME} PUBLIC ./ )
-`
-module.exports = async(args, zeusbox) => {
-    if (zeusbox.install && zeusbox.install.contracts) {
-        await Promise.all(Object.keys(zeusbox.install.contracts).map(async contract => {
-            console.log(emojMap.eight_spoked_asterisk + `Configuring ${contract.green}`);
+`;
+module.exports = async (args, zeusbox) => {
+  if (zeusbox.install && zeusbox.install.contracts) {
+    await Promise.all(Object.keys(zeusbox.install.contracts).map(async contract => {
+      console.log(emojMap.eight_spoked_asterisk + `Configuring ${contract.green}`);
 
-            var fileName = './contracts/eos/CMakeLists.txt';
-            var cmakelists = "";
-            if (fs.existsSync(fileName))
-                cmakelists = fs.readFileSync(fileName).toString();
-            cmakelists += `\nExternalProject_Add(
+      var fileName = './contracts/eos/CMakeLists.txt';
+      var cmakelists = '';
+      if (fs.existsSync(fileName)) { cmakelists = fs.readFileSync(fileName).toString(); }
+      cmakelists += `\nExternalProject_Add(
                ${contract}
                SOURCE_DIR ${contract}
                BINARY_DIR ${contract}
@@ -31,12 +30,11 @@ module.exports = async(args, zeusbox) => {
                TEST_COMMAND ""
                INSTALL_COMMAND ""
                BUILD_ALWAYS 1
-            )\n`
-            fs.writeFileSync(fileName, cmakelists);
-            var contractDir = path.join('./contracts/eos/', contract);
-            var contractMakeFile = path.join(contractDir, 'CMakeLists.txt');
-            if (fs.existsSync(contractDir) && !fs.existsSync(contractMakeFile))
-                fs.writeFileSync(contractMakeFile, CMakeLists);
-        }));
-    }
-}
+            )\n`;
+      fs.writeFileSync(fileName, cmakelists);
+      var contractDir = path.join('./contracts/eos/', contract);
+      var contractMakeFile = path.join(contractDir, 'CMakeLists.txt');
+      if (fs.existsSync(contractDir) && !fs.existsSync(contractMakeFile)) { fs.writeFileSync(contractMakeFile, CMakeLists); }
+    }));
+  }
+};
